move login redux state into its own reducer
